+++ src/core/Table/Filters/index.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export { FilterWrapper } from './FilterWrapper/FilterWrapper';
+export { FilterWrapper } from './FilterWrapper';
 
 export type { TableFilterProps, TableFilterValue } from './types';
